+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file contains the full list of changes for every version of Cherry Doom.
 
 - _Dithered Lighting_ setting
 - _Pulsating Message Display_ setting
+- _[Weapon] Switch Speed_ setting
 
 ### Changed